apply coderabbit fixes
