+++ apps/desktop/i18n/es.json
@@ -293,6 +293,7 @@
         "checkpoint": {
             "created": "{axis}-punto de control creado en {steps} pasos",
             "delete": "Eliminar",
+            "deleteTooltip": "Eliminar este punto de control del campo",
             "deleted": "{axis}-punto de control eliminado en {steps} pasos - \"{name}\"",
             "fieldLabel": "Etiqueta de Campo",
             "fieldLabelTooltip": "La etiqueta para aparecer en el campo (es decir, las marcas de yarda)",
@@ -417,6 +418,7 @@
             "leftAbbreviation": "P. ej. \"S1\", \"IP\" o \"ED\" (abreviatura de \"Lado 1\", \"Izquierda del público\" o \"Escenario derecho\")",
             "measurementSystem": "La unidad de medida para definir el tamaño del paso. Se puede ir hacia adelante y hacia atrás",
             "resetTheme": "Restablecer el tema a los valores predeterminados",
+            "resetToDefault": "Restablecer al predeterminado",
             "rightAbbreviation": "P. ej. \"S2\", \"AR\" o \"IS\" (abreviatura de \"Lado 2\", \"Derecha del público\" o \"Izquierda del escenario\")",
             "stepsFromFrontToAwayLabelBottom": "Número de pasos desde la línea lateral delantera hasta el exterior del número visitante (más cerca de la línea de fondo)",
             "stepsFromFrontToAwayLabelTop": "Número de pasos desde la línea lateral delantera hasta el interior del número visitante (más cerca del centro del campo)",
