+++ apps/desktop/src/components/field/customizer/CheckpointsTab.tsx
@@ -1,10 +1,22 @@
-import { FieldProperties } from "@openmarch/core";
+import { Checkpoint, FieldProperties } from "@openmarch/core";
 import { TabContent, Button, Switch } from "@openmarch/ui";
 import { T, useTolgee } from "@tolgee/react";
-import FormField from "../../ui/FormField";
-import { CheckpointsTabProps } from "./types";
+import FormField from "@/components/ui/FormField";
 import { CheckpointEditor } from "./CheckpointEditor";
 import { inputClassname } from "./utils";
+
+export interface CheckpointsTabProps {
+    currentFieldProperties: FieldProperties;
+    updateFieldProperties: (props: FieldProperties) => void;
+    updateCheckpoint: (args: {
+        oldCheckpoint: Checkpoint;
+        newCheckpoint: Checkpoint;
+        axis: "x" | "y";
+    }) => void;
+    deleteCheckpoint: (checkpoint: Checkpoint) => void;
+    addCheckpoint: (axis: "x" | "y") => void;
+    sorter: (a: Checkpoint, b: Checkpoint) => number;
+}
 
 export function CheckpointsTab({
     currentFieldProperties,
